docs(README): Clean up code blocks
Removes prompt character in code blocks so when it's rendered on GitHub, the
copy button is useful for quick pasting in the terminal.

Also updates the go installation commands for the wego dependency setup -
`go get` is no longer supported outside a module, and the `go install` command
needs to reference a specific git reference.