Fix wttrin link to reflect new owner of Emacs Melpa package.
The Location for Emacs wttrin package listed in the integration.md is no longer correct.

I have adopted the Emacs wttrin package, provided fixes and enhancements. and relocated it to it's new location in this PR. Please update so users can find the current working package as published on Melpa.